minor type cleanups

+++ inspect.c
@@ -16,30 +16,23 @@
 /* inspects loaded gif context and prints details */
 gif_error inspect_gif(const gif_context* context)
 {
-    /* validate context */
     if (!context)
         return GIF_ERROR_INVALID_FORMAT;
 
     /* calculate duration statistics */
-    
     u32 total_duration_ms = 0;
-    u32 min_delay = 0xFFFFFFFF;     /* maximum of u32 */
+    u32 min_delay = 0xFFFFFFFF;
     u32 max_delay = 0;
 
-    for (usize i = 0; i < (usize)context->frames; i++)
+    for (s32 i = 0; i < context->frames; i++)
     {
         /* delay is in ms */
         u32 delay = (u32)context->delays[i];
 
-        /* accumulate total duration */
         total_duration_ms += delay;
 
-        /* check min */
-        if (delay < min_delay)
-            min_delay = delay;
-        /* check max */
-        if (delay > max_delay)
-            max_delay = delay;
+        if (delay < min_delay) min_delay = delay;
+        if (delay > max_delay) max_delay = delay;
     }
 
     f32 duration_sec = (f32)total_duration_ms / 1000.0f;
@@ -58,11 +51,10 @@ gif_error inspect_gif(const gif_context* context)
     printf("  Maximum:      %u ms\n", max_delay);
     printf("=============================\n");
     
-    /* if there are few frames, list them specifically */
     if (context->frames <= 10)
     {
         printf("Frame List:\n");
-        for (int i = 0; i < context->frames; i++)
+        for (s32 i = 0; i < context->frames; i++)
         {
             printf("  Frame %02d: %d ms\n", i, context->delays[i]);
         }
